fixed the summary response dto, removed redunted statistics

+++ server/allTheMethods/src/main/java/com/example/allTheMethods/dto/response/SummaryResponseDto.java
@@ -4,6 +4,6 @@ public record SummeryResponseDto(
         int countTotalProblems,
         int countTotalAchievements,
         int countTotalSubmissions,
-        int userAcceptanceRate
+        float userAcceptanceRate
 ) {
 }
